initial

+++ gfs2/data_model.go
@@ -5,6 +5,6 @@ type ClusterMetric map[string]FSMetric
 type FSMetric map[string]Metric
 
 type Metric struct {
-	Value uint16
+	Value uint64
 	State string
 }
